Upload practice projects with mentor's suggestion
