--- jzp-code/boot-demo/src/main/java/com/tutorial/bootDemo/BootDemoApplication.java
@@ -1,12 +1,0 @@
-package com.tutorial.bootDemo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BootDemoApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(BootDemoApplication.class, args);
-    }
-
-}
