--- jzp-code/boot-demo/src/main/java/com/tutorial/bootDemo/TestController.java
@@ -1,15 +1,0 @@
-package com.tutorial.bootDemo;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-public class TestController {
-
-    @GetMapping("/hello")
-    public List<String> hello() {
-        return List.of("Hello", "world");
-    }
-}
