--- jzp-code/boot-demo/src/main/java/com/tutorial/bootDemo/service/StudentServiceImpl.java
@@ -1,56 +1,0 @@
-package com.tutorial.bootDemo.service;
-
-import com.tutorial.bootDemo.converter.StudentConverter;
-import com.tutorial.bootDemo.dao.Student;
-import com.tutorial.bootDemo.dao.StudentRepository;
-import com.tutorial.bootDemo.dto.StudentDTO;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.CollectionUtils;
-import org.springframework.util.StringUtils;
-
-import java.util.List;
-
-@Service
-public class StudentServiceImpl implements StudentService {
-
-    @Autowired
-    private StudentRepository studentRepository;
-
-    @Override
-    public StudentDTO getStudentById(long id) {
-        Student student = studentRepository.findById(id).orElseThrow(RuntimeException::new);
-        return StudentConverter.convertStudent(student);
-    }
-
-    @Override
-    public Long addNewStudent(StudentDTO studentDTO) {
-        List<Student> studentList = studentRepository.findByEmail(studentDTO.getEmail());
-        if (!CollectionUtils.isEmpty(studentList)) {
-            throw new IllegalStateException("email:" + studentDTO.getEmail() + " has been taken");
-        }
-        Student student = studentRepository.save(StudentConverter.convertStudent(studentDTO));
-        return student.getId();
-    }
-
-    @Override
-    public void deleteStudentById(long id) {
-        studentRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("id" + id + "doesn't exist!"));
-        studentRepository.deleteById(id);
-    }
-
-    @Override
-    @Transactional
-    public StudentDTO updateStudentById(long id, String name, String email) {
-        Student studentInDB = studentRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("id" + id + "doesn't exist!"));
-        if (StringUtils.hasLength(name) && !studentInDB.getName().equals(name)) {
-            studentInDB.setName(name);
-        }
-        if (StringUtils.hasLength(email) && !studentInDB.getEmail().equals(email)) {
-            studentInDB.setEmail(email);
-        }
-        Student student = studentRepository.save(studentInDB);
-        return StudentConverter.convertStudent(student);
-    }
-}
