+++ jzp-code/practice-projects/docker_test/Dockerfile
