--- jzp-code/spring_test/.idea/.gitignore
@@ -1,3 +1,0 @@
-# 默认忽略的文件
-/shelf/
-/workspace.xml
