--- jzp-code/spring_test/src/main/java/MyApplication.java
@@ -1,21 +1,0 @@
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@SpringBootApplication
-public class MyApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MyApplication.class, args);
-    }
-
-    @RestController
-    public static class HelloController {
-
-        @GetMapping("/")
-        public String hello() {
-            return "Hello, World!";
-        }
-    }
-}
